EduModel 做完了绩点

+++ database/migrations/2017_03_16_020332_create_edu_grades_table.php
@@ -23,7 +23,7 @@ class CreateEduGradesTable extends Migration
             $table->float('grade');
             $table->unsignedTinyInteger('state')->default(0)->comment('类型,0正常,1补考,2重修,3免修');
             $table->unsignedInteger('teacher_id');
-            $table->unsignedInteger('credit')->comment('所占学分');
+            $table->float('credit')->comment('所占学分');
             $table->foreign('course_id')->references('id')->on('edu_courses');
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('teacher_id')->references('id')->on('edu_teachers');
